Added support for scrolling using the trackball.
Currently scrolling is selected in shift mode or with the shift
modifier.  This sets the y-motion of the trackball to scroll with a
lower resolution than pointer movement; currently set to the lowest
possible with the ADNS-9800 which is 50 cpi.

If a lower resolution is required then the movement of the trackball
will need to be accumulated in the TrackBall class so that the
resolution may be reduced before sending.